use `boost::math::ccmath::sqrt` instead of `std::sqrt`
