+++ Cargo.toml
@@ -9,6 +9,11 @@ include = [
   "wrapper.cpp",
   #
   "subprojects/boost_math/math/include/boost/math/ccmath/detail",
+  "subprojects/boost_math/math/include/boost/math/ccmath/abs.hpp",
+  "subprojects/boost_math/math/include/boost/math/ccmath/isnan.hpp",
+  "subprojects/boost_math/math/include/boost/math/ccmath/isinf.hpp",
+  "subprojects/boost_math/math/include/boost/math/ccmath/sqrt.hpp",
+  #
   "subprojects/boost_math/math/include/boost/math/policies",
   #
   "subprojects/boost_math/math/include/boost/math/special_functions",
