+++ src/math/special_functions/mod.rs
@@ -13,6 +13,7 @@ pub mod legendre;
 pub mod logsumexp;
 pub mod polygamma;
 pub mod prime;
-pub mod sqrt;
+pub mod rsqrt;
+pub mod sqrt1pm1;
 pub mod trigamma;
 pub mod zeta;
